[refactor] - 상품 등록 Api 수정

+++ src/main/java/com/wanted/ecommerce/product/dto/request/ProductCreateRequest.java
@@ -25,5 +25,5 @@ public class ProductCreateRequest {
     private List<ProductCategoryRequest> categories;
     private List<ProductOptionGroupRequest> optionGroups;
     private List<ProductImageRequest> images;
-    private List<Integer> tags;
+    private List<Long> tags;
 }
